requirements.txt
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorlog==6.10.1
contourpy==1.3.3
coverage==7.12.0
cryptography==46.0.3
cycler==0.12.1
databricks-sdk==0.73.0
deap==1.4.3
docker==7.1.0
fastapi==0.121.3
Flask==3.1.2
flask-cors==6.0.1
fonttools==4.60.1
-e git+https://github.com/ibrah5em/ga-optimized-trees.git@adb6f5be77a7bcf484648e196837d21286b3bcee#egg=ga_optimized_trees
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.43.0
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
graphviz==0.21
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
huey==2.5.4
idna==3.11
ImageIO==2.37.2
importlib_metadata==8.7.0
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
lazy_loader==0.4
lime==0.2.0.1
llvmlite==0.45.1
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.7
mlflow==3.6.0
mlflow-skinny==3.6.0
mlflow-tracing==3.6.0
networkx==3.5
numba==0.62.1
numpy==2.3.5
nvidia-nccl-cu12==2.28.9
opentelemetry-api==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
optuna==4.6.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pluggy==1.6.0
protobuf==6.33.1
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
pytest==9.0.1
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
shap==0.50.0
six==1.17.0
slicer==0.0.8
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlparse==0.5.3
starlette==0.50.0
threadpoolctl==3.6.0
tifffile==2025.10.16
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
Werkzeug==3.1.3
xgboost==3.1.2
zipp==3.23.0
